Made updater work with 2.80
