+++ __init__.py
@@ -28,6 +28,7 @@ from . import facerig
 from . import humanoid, animationengine, proxyengine
 from . import utils
 from . import algorithms
+from . import preferences
 from . import addon_updater_ops
 
 logger = logging.getLogger(__name__)
@@ -1806,69 +1807,6 @@ class LoadTemplate(bpy.types.Operator):
         return {'FINISHED'}
 
 
-# demo bare-bones preferences
-@addon_updater_ops.make_annotations
-class MBPreferences(bpy.types.AddonPreferences):
-    '''
-    Updater preferences
-    '''
-    bl_idname = __package__
-    # addon updater preferences
-    auto_check_update = bpy.props.BoolProperty(
-        name="Auto-check for Update",
-        description="If enabled, auto-check for updates using an interval",
-        default=False,
-    )
-    updater_intrval_months = bpy.props.IntProperty(
-        name='Months',
-        description="Number of months between checking for updates",
-        default=0,
-        min=0
-    )
-    updater_intrval_days = bpy.props.IntProperty(
-        name='Days',
-        description="Number of days between checking for updates",
-        default=7,
-        min=0,
-        max=31
-    )
-    updater_intrval_hours = bpy.props.IntProperty(
-        name='Hours',
-        description="Number of hours between checking for updates",
-        default=0,
-        min=0,
-        max=23
-    )
-    updater_intrval_minutes = bpy.props.IntProperty(
-        name='Minutes',
-        description="Number of minutes between checking for updates",
-        default=0,
-        min=0,
-        max=59
-    )
-
-    def draw(self, context):
-        layout = self.layout
-        # col = layout.column() # works best if a column, or even just self.layout
-        mainrow = layout.row()
-        col = mainrow.column()
-
-        # updater draw function
-        # could also pass in col as third arg
-        addon_updater_ops.update_settings_ui(self, context)
-
-        # Alternate draw function, which is more condensed and can be
-        # placed within an existing draw function. Only contains:
-        #   1) check for update/update now buttons
-        #   2) toggle for auto-check (interval will be equal to what is set above)
-        # addon_updater_ops.update_settings_ui_condensed(self, context, col)
-
-        # Adding another column to help show the above condensed ui as one column
-        # col = mainrow.column()
-        # col.scale_y = 2
-        # col.operator("wm.url_open","Open webpage ").url=addon_updater_ops.updater.website
-
-
 class VIEW3D_PT_tools_ManuelbastioniLAB(bpy.types.Panel):
     bl_label = "MB-Lab {0}.{1}.{2}".format(bl_info["version"][0], bl_info["version"][1], bl_info["version"][2])
     bl_idname = "OBJECT_PT_characters01"
@@ -2340,10 +2278,9 @@ classes = (
     CreateFaceRig,
     DeleteFaceRig,
     LoadTemplate,
-    MBPreferences,
+    preferences.MBPreferences,
     VIEW3D_PT_tools_ManuelbastioniLAB,
 )
-
 
 def register():
     # addon updater code and configurations
@@ -2353,7 +2290,6 @@ def register():
 
     # register the example panel, to show updater buttons
     for cls in classes:
-        addon_updater_ops.make_annotations(cls)  # to avoid blender 2.8 warnings
         bpy.utils.register_class(cls)
